feat(diagnose): FritzBox reconnect action + fix-buttons in onboarding wizard
After `configure_fritzbox` sets DHCP option 6 to point at AdGuard,
clients on existing leases keep their old DNS until the lease renews
(FritzBox default: 24 h). A WAN reconnect causes the box to re-read its
DNS config and re-publish DHCP state — the single quickest way to make
the new setting take effect across the LAN without rebooting devices.

Adds:
- `reconnectFritzBox()` helper in lib/router/dnsConfig.ts. Calls TR-064
  `ForceTermination` on WANIPConnection, falls back to WANPPPConnection
  for DSL/PPPoE installs. Shares the `RouterCallResult` shape with
  `setFritzBoxDhcpDns` so probe-action callers don't branch by helper.
- `reconnect_fritzbox` action on the `router_dns_not_pointing` probe.
  Destructive (brief WAN outage, potential new public IP on DSL).
- Shared `<DiagnoseProbeList>` component — extracted from
  `SelfDiagnoseSection` so the OnboardingWizard's post-install
  diagnose panel renders the same fix-buttons (incl. the new
  reconnect button) instead of just status + hint text. Compact mode
  hides ok-rows for the wizard's tight summary view.

`SelfDiagnoseSection` shrinks to its outer chrome + the shared list;
behavior is unchanged. Wizard auto-runs diagnose on phase=done and now
captures `node` so dispatch targets the right host.
